+++ packages/dx-react-grid/docs/reference/integrated-filtering.md
