+++ packages/dx-react-grid/docs/reference/integrated-grouping.md
