+++ packages/dx-react-grid/docs/reference/integrated-paging.md
