+++ packages/dx-react-grid/docs/reference/integrated-selection.md
