+++ packages/dx-react-grid/docs/reference/integrated-sorting.md
