+++ packages/dx-react-grid/src/plugins/integrated-filtering.jsx
