+++ packages/dx-react-grid/src/plugins/integrated-grouping.jsx
