+++ packages/dx-react-grid/src/plugins/integrated-paging.jsx
