+++ packages/dx-react-grid/src/plugins/integrated-selection.jsx
