+++ packages/dx-react-grid/src/plugins/integrated-sorting.jsx
